@@ -9,12 +9,9 @@
   },
   "scripts": {
     "analyze": "source-map-explorer build/static/js/main.*",
-    "start": "npm-run-all -p watch:css start:js",
-    "start:js": "react-scripts start",
+    "start": "react-scripts start",
     "start:prod": "sh -ac '. .env.prod; npm start'",
-    "build": "npm run build:css && react-scripts build",
-    "build:css": "node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/",
-    "watch:css": "npm run build:css && node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/ --watch --recursive",
+    "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom --coverage",
     "test:watch": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
@@ -30,12 +27,11 @@
     "connected-react-router": "^4.4.0",
     "copy-to-clipboard": "^3.0.8",
     "dayjs": "^1.6.10",
-    "firebase": "^5.3.0",
+    "firebase": "^5.5.3",
     "history": "^4.7.2",
     "is-url-external": "^1.0.3",
     "lodash": "^4.17.10",
-    "node-sass-chokidar": "1.3.0",
-    "npm-run-all": "^4.1.3",
+    "node-sass": "^4.9.3",
     "pluralize": "^7.0.0",
     "prop-types": "^15.6.2",
     "react": "^16.4.1",
@@ -63,17 +59,17 @@
     "babel-eslint": "^9.0.0",
     "enzyme": "^3.1.1",
     "enzyme-adapter-react-16": "^1.0.4",
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-config-prettier": "^2.6.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-prettier": "^2.6.1",
-    "eslint-plugin-react": "^7.10.0",
-    "firebase-tools": "^4.0.0",
-    "prettier": "^1.13.7",
-    "react-scripts": "^1.0.1",
+    "eslint": "^5.6.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-prettier": "^3.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-prettier": "^2.7.0",
+    "eslint-plugin-react": "^7.11.1",
+    "firebase-tools": "^5.0.1",
+    "prettier": "^1.14.3",
+    "react-scripts": "2.0.4",
     "react-test-renderer": "^16.4.1",
     "source-map-explorer": "^1.6.0"
   },
@@ -83,5 +79,11 @@
       "!node_modules",
       "!src/components/**/index.js"
     ]
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
