@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
@@ -6,20 +6,20 @@ import Header, { HeaderType } from 'primitives/text/header';
 import FlexContainer from 'primitives/container/flex-container';
 import Loading from 'primitives/regions/loading';
 
-import './style.css';
+import './style.scss';
 
 export default function EmptyOverview({ children, isInitialized }) {
   let body = <Loading />;
   if (isInitialized) {
     body = (
-      <Fragment>
+      <>
         <Header type={HeaderType.header2} className="EmptyOverview-Header1">
           <FormattedMessage id="misc.entityOverview" />
         </Header>
         <Header type={HeaderType.header3} className="EmptyOverview-Header2">
           {children || <FormattedMessage id="misc.selectEntity" />}
         </Header>
-      </Fragment>
+      </>
     );
   }
   return (
