@@ -9,7 +9,7 @@ import ExportTypes from 'constants/export-types';
 import { createJSONDownload, createImageDownlaod } from 'utils/export';
 import { mapValues, omit } from 'constants/lodash';
 
-import './style.css';
+import './style.scss';
 
 export default function ExportModal({
   isExportOpen,
@@ -30,7 +30,8 @@ export default function ExportModal({
         ),
         sector.name,
       );
-    } else if (exportType === ExportTypes.image.key) {
+    }
+    if (exportType === ExportTypes.image.key) {
       closeExport();
       return createImageDownlaod('hex-map');
     }
