@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
@@ -8,7 +8,7 @@ import BasicLink from 'primitives/other/basic-link';
 
 import { RotateCcw } from 'constants/icons';
 
-import './style.css';
+import './style.scss';
 
 const titledItem = (key, description, title) => {
   let titleElement;
@@ -83,7 +83,7 @@ export default function FactionAttributes({ faction, attributes, className }) {
           {faction.tags.map(tag =>
             titledItem(
               tag,
-              <Fragment>
+              <>
                 <FormattedMessage id={`faction.tags.${tag}.description`} />
                 <FlexContainer className="FactionAttributes-TagEffect">
                   <span className="FactionAttributes-Effect">
@@ -91,7 +91,7 @@ export default function FactionAttributes({ faction, attributes, className }) {
                   </span>
                   <FormattedMessage id={`faction.tags.${tag}.effect`} />
                 </FlexContainer>
-              </Fragment>,
+              </>,
               <FormattedMessage id={`faction.tags.${tag}`} />,
             ),
           )}
