@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Measure from 'react-measure';
@@ -21,7 +21,7 @@ import {
 } from 'constants/icons';
 import { isArray } from 'constants/lodash';
 
-import './style.css';
+import './style.scss';
 
 const formatOptionalMessage = (intl, key, builder) => {
   const id = builder(key);
@@ -98,10 +98,10 @@ const buildFactionTableColumns = ({ intl, windowWidth, sector }) => {
         let incomeElement;
         if (income) {
           incomeElement = (
-            <Fragment>
+            <>
               <RotateCcw size={12} className="FactionTable-Income" />
               {income}
-            </Fragment>
+            </>
           );
         }
         const convertedBalance = typeof balance === 'string' ? 0 : balance || 0;
@@ -239,7 +239,8 @@ export default class FactionTable extends Component {
 
     if (isLoading) {
       return <Loading />;
-    } else if (!isSaved) {
+    }
+    if (!isSaved) {
       return <FactionNotSaved />;
     }
 
