@@ -1,20 +1,23 @@
-import React, { Component, Fragment } from 'react';
+import { Component } from 'react';
 import PropTypes from 'prop-types';
 
 export default class InitWrapper extends Component {
-  componentWillMount() {
-    this.props.initialize(this.props.location);
+  static propTypes = {
+    children: PropTypes.node.isRequired,
+    initialize: PropTypes.func.isRequired,
+    location: PropTypes.shape({
+      pathname: PropTypes.string.isRequired,
+    }).isRequired,
+  };
+
+  constructor(props) {
+    super(props);
+
+    props.initialize(props.location);
   }
 
   render() {
-    return <Fragment>{this.props.children}</Fragment>;
+    const { children } = this.props;
+    return children;
   }
 }
-
-InitWrapper.propTypes = {
-  children: PropTypes.node.isRequired,
-  initialize: PropTypes.func.isRequired,
-  location: PropTypes.shape({
-    pathname: PropTypes.string.isRequired,
-  }).isRequired,
-};
