@@ -6,7 +6,7 @@ import { FormattedMessage, intlShape } from 'react-intl';
 import FlexContainer from 'primitives/container/flex-container';
 import Input from 'primitives/form/input';
 
-import './style.css';
+import './style.scss';
 
 export default function NoteSidebar({
   isSidebarEditActive,
