@@ -1,27 +1,22 @@
 import { connect } from 'react-redux';
 import { injectIntl } from 'react-intl';
+import { createStructuredSelector } from 'reselect';
 
 import { cancelTopLevelEntityCreate } from 'store/actions/sector.actions';
 import { generateEntity } from 'store/actions/entity.actions';
-import { getCurrentTopLevelEntities } from 'store/selectors/entity.selectors';
+import { topLevelEntityModalOpen } from 'store/selectors/entity.selectors';
 import {
   topLevelKeySelector,
   currentSectorSelector,
 } from 'store/selectors/base.selectors';
-import { getTopLevelEntity } from 'utils/entity';
 
 import TopLevelEntityModal from './top-level-entity-modal';
 
-const mapStateToProps = state => {
-  const topLevelKey = topLevelKeySelector(state);
-  return {
-    topLevelKey,
-    currentSector: currentSectorSelector(state),
-    isOpen:
-      !!topLevelKey &&
-      !!getTopLevelEntity(getCurrentTopLevelEntities(state), topLevelKey),
-  };
-};
+const mapStateToProps = createStructuredSelector({
+  topLevelKey: topLevelKeySelector,
+  currentSector: currentSectorSelector,
+  isOpen: topLevelEntityModalOpen,
+});
 
 const mapDispatchToProps = (dispatch, props) => ({
   cancelTopLevelEntityCreate: () => dispatch(cancelTopLevelEntityCreate()),
