@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-import './style.css';
+import './style.scss';
 
 function AbsoluteContainer({ forwardedRef, children, className, ...props }) {
   return (
