@@ -4,7 +4,7 @@ import classNames from 'classnames';
 
 import Spinner from 'primitives/other/spinner';
 
-import './style.css';
+import './style.scss';
 
 export default function Button({
   children,
@@ -24,6 +24,7 @@ export default function Button({
   return (
     <button
       {...rest}
+      type="button"
       className={classNames(className, {
         Button: !minimal,
         'Button-Minimal': minimal,
